+++ backend/TimeTile/TimeTile.Core/Common/Interfaces/Services/IFileService.cs
@@ -1,4 +1,5 @@
-﻿using TimeTile.Core.Common.UnifiedResponse;
+﻿using Microsoft.AspNetCore.Http;
+using TimeTile.Core.Common.UnifiedResponse;
 using TimeTile.Core.Enums;
 
 namespace TimeTile.Core.Common.Interfaces.Services;
@@ -8,4 +9,5 @@ public interface IFileService
     Task<int> SaveFile(Stream fileStream, string fileName, CancellationToken cancellationToken);
     Task DeleteFilePhysically(int id, CancellationToken cancellationToken);
     string GetContentType(FileExtension extension);
+    Task<List<int>> SaveFiles(IFormFileCollection files, CancellationToken cancellationToken);
 }
